# api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Poker Arena API",
    "summary": "Texas Hold'em No-Limit 6-max contra IAs de níveis configuráveis.",
...
    "contact": {
      "name": "Poker Arena",
      "url": "http://localhost:5173/"
    },
    "license": {
      "name": "Uso educacional (feira de ciências)"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/copilot": {
      "post": {
        "tags": [
          "Copiloto"
        ],
        "summary": "Copiloto: revisar um spot (pós-jogo, offline)",
        "description": "Analisa um SPOT que você descreve (suas cartas, board, pote, preço, posição)\ne devolve a leitura completa: equity real, pot odds, EV, MDF, outs, a nut,\ntextura, blockers, o veredito de CADA jogada (boa/arriscada/ruim + por quê) e o\nconselho das 5 IAs. É o painel 'Sua jogada' aplicado a qualquer situação — 100%\noffline, sem tocar em site nenhum. Ideal pra ESTUDAR/revisar mãos depois do jogo.",
        "operationId": "copilot_copilot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopilotResponse"
                }
              }
            }
          },
          "400": {
            "description": "Spot inválido (cartas repetidas, quantidade errada, valores inválidos)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/from-image": {
      "post": {
        "tags": [
          "Copiloto"
        ],
        "summary": "Copiloto: ler uma IMAGEM da mesa e decidir (visão computacional)",
        "description": "VISÃO → estado → sanity → Copiloto. A imagem é lida (cartas + pote) pela\nvisão; o estado passa pelo sanity-check de regras (a rede de segurança); se\npassar, o Copiloto decide. Se a leitura for implausível, ABSTÉM (não decide\ncom lixo). Pós-jogo/estudo, offline — não lê tela de jogo ao vivo.",
        "operationId": "from_image_copilot_from_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_from_image_copilot_from_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FromImageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Imagem inválida."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/copilot/review-hand": {
      "post": {
        "tags": [
          "Copiloto"
        ],
        "summary": "Copiloto: revisar uma MÃO inteira (cole o histórico PHH)",
        "description": "Cole o histórico da mão (formato PHH — o texto que o jogo exporta ao FIM da\nmão) e escolha qual jogador é você. O copiloto reproduz a mão e revisa CADA\ndecisão sua, comparando com a recomendação. É estudo pós-jogo (como rever um\nPGN de xadrez), 100% offline — não lê tela de jogo ao vivo.",
        "operationId": "review_hand_endpoint_copilot_review_hand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandReviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "Histórico inválido, jogador inexistente ou sem decisões suas."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/games/{game_id}": {
      "get": {
        "tags": [
          "Auditoria"
        ],
        "summary": "Replay completo de uma partida",
        "description": "Devolve o log completo de uma partida: cada mão com as ações, o **raciocínio\nde cada bot** no momento (glass-box), o board, os vencedores e o saldo em fichas.",
        "operationId": "game_games__game_id__get",
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Game Games  Game Id  Get"
                }
              }
            }
          },
          "404": {
            "description": "Partida não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables": {
      "post": {
        "tags": [
          "Mesa"
        ],
        "summary": "Criar uma mesa (nova partida)",
        "description": "Cria uma partida com os bots, blinds, formato (cash/torneio) e modo (play/watch)\nescolhidos. **Guarde o `table_id` da resposta** — ele identifica a mesa em todas\nas chamadas seguintes.",
        "operationId": "create_table_tables_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Estado inicial da mesa, já com a 1ª mão distribuída.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Nível de bot inválido ou Expert indisponível."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_id}": {
      "get": {
        "tags": [
          "Mesa"
        ],
        "summary": "Ler o estado atual da mesa",
        "description": "Devolve o estado completo e atual da mesa (fase, cartas, pote, cadeiras,\njogadas válidas, análises). É o jeito de 'dar refresh' sem alterar nada.",
        "operationId": "get_table_tables__table_id__get",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Mesa não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_id}/actions": {
      "post": {
        "tags": [
          "Jogada"
        ],
        "summary": "Fazer a sua jogada (turno do humano)",
        "description": "Aplica a sua jogada quando `phase == human_turn`. Só valem as ações listadas\nem `legal.actions`; para `raise`, o `amount` é o valor TOTAL (entre `min_raise_to`\ne `max_raise_to`). Os bots jogam sozinhos em seguida, até voltar a ser a sua vez.",
        "operationId": "act_tables__table_id__actions_post",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Jogada ilegal (não está em `legal.actions` ou valor fora do permitido)."
          },
          "404": {
            "description": "Mesa não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_id}/next-hand": {
      "post": {
        "tags": [
          "Jogada"
        ],
        "summary": "Começar a próxima mão",
        "description": "Distribui uma nova mão. Só é válido quando `phase == hand_over`.",
        "operationId": "next_hand_tables__table_id__next_hand_post",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "A mão atual ainda não terminou (`phase` != `hand_over`)."
          },
          "404": {
            "description": "Mesa não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_id}/step": {
      "post": {
        "tags": [
          "Jogada"
        ],
        "summary": "Avançar uma jogada de bot (Modo Laboratório)",
        "description": "No modo `watch`, avança UMA jogada de bot por vez — é assim que você acompanha\na partida lance a lance, vendo o raciocínio de cada IA. Válido só em `bot_turn`.",
        "operationId": "step_tables__table_id__step_post",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Não há jogada de bot pendente (`phase` != `bot_turn`)."
          },
          "404": {
            "description": "Mesa não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tables/{table_id}/players": {
      "post": {
        "tags": [
          "Jogadores"
        ],
        "summary": "Sentar um novo bot na mesa",
        "description": "Adiciona um bot do nível escolhido. A mudança vale **a partir da próxima mão**\n(a mão atual termina normalmente). Mesa 6-max.",
        "operationId": "add_player_tables__table_id__players_post",
        "parameters": [
          {
            "name": "table_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Table Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPlayerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableStateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Mesa cheia (máx. 6) ou nível inválido."
          },
          "404": {
            "description": "Mesa não encontrada."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CopilotRequest": {
        "properties": {
          "hole": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hole",
            "description": "Suas 2 cartas (ex.: ['As','Kh'])."
          },
          "board": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Board",
            "description": "Cartas comunitárias: 0 (pré-flop), 3 (flop), 4 (turn) ou 5 (river).",
            "default": []
          },
          "pot": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pot",
            "description": "Fichas no pote (antes de você pagar)."
          },
          "to_call": {
            "type": "integer",
            "minimum": 0.0,
            "title": "To Call",
            "description": "Quanto custa pagar (0 se você pode passar).",
            "default": 0
          },
          "my_stack": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "My Stack",
            "description": "Suas fichas."
          },
          "num_opponents": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Num Opponents",
            "description": "Quantos oponentes ativos na mão.",
            "default": 1
          },
          "in_position": {
            "type": "boolean",
            "title": "In Position",
            "description": "Você age por último (em posição)? Afeta a realização da equity.",
            "default": true
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position",
            "description": "Posição na mesa: SB/BB/UTG/UTG+1/MP1/MP2/DJ/HJ/BTN. Se informada, o copiloto segue a regra oficial (mais apertado em posição cedo)."
          },
          "big_blind": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Big Blind",
            "description": "Big blind (só para dimensionar o aumento mínimo).",
            "default": 20
          }
        },
        "type": "object",
        "required": [
          "hole",
          "pot",
          "my_stack"
        ],
        "title": "CopilotRequest",
        "description": "Um SPOT descrito pelo usuário para o Copiloto revisar (pós-jogo, offline)."
      },
...
    }
  },
...
}